+++ packages/backend-core/src/cache/index.ts
@@ -1,8 +1,8 @@
+export * as docWritethrough from "./docWritethrough"
+export * from "./generic"
 export * as generic from "./generic"
-export * as user from "./user"
-export * as app from "./appMetadata"
-export * as writethrough from "./writethrough"
 export * as invite from "./invite"
 export * as passwordReset from "./passwordReset"
-export * from "./generic"
-export * as docWritethrough from "./docWritethrough"
+export * as user from "./user"
+export * as workspace from "./workspaceMetadata"
+export * as writethrough from "./writethrough"
