+++ packages/server/src/middleware/recaptcha.ts
@@ -23,8 +23,8 @@ const middleware = (async (ctx: UserCtx, next: Next) => {
   ) {
     return next()
   }
-  const app = await cache.app.getWorkspaceMetadata(workspaceId)
-  if ("state" in app && app.state === cache.app.WorkspaceState.INVALID) {
+  const app = await cache.workspace.getWorkspaceMetadata(workspaceId)
+  if ("state" in app && app.state === cache.workspace.WorkspaceState.INVALID) {
     throw new Error("App not found")
   }
   if ((app as Workspace).recaptchaEnabled) {
