+++ packages/server/src/sdk/workspace/backups/errors.ts
@@ -10,5 +10,5 @@ export async function clearErrors(backupId?: string) {
     delete metadata.backupErrors[backupId]
   }
   await database.put(metadata)
-  await cache.app.invalidateWorkspaceMetadata(metadata.appId, metadata)
+  await cache.workspace.invalidateWorkspaceMetadata(metadata.appId, metadata)
 }
